fixed erroneous matches in tarballs
Matcher was searching for files ending with binary listed in distributions

thus, `_rg` would match `rg` specification, and since the former appears after the latter in the tarball, `_ rg` would be installed as the binary (but the file contains the completion).

To solve this, added support for regexps in distribution binaries specification.

Fixes #239